+++ apps/e2e/ag-grid-storybook-e2e/src/support/commands.ts
@@ -2,23 +2,6 @@
 declare namespace Cypress {
   // eslint-disable-next-line @typescript-eslint/no-unused-vars
   interface Chainable<Subject> {
-    waitForFonts(...fonts: string[]): Chainable<void>;
-
-    /**
-     * Wait for Font Awesome to be loaded.
-     */
-    waitForFontAwesome(): Chainable<void>;
-
-    /**
-     * Wait for BLKB Sans to be loaded.
-     */
-    waitForBlackbaudSans(): Chainable<void>;
-
-    /**
-     * Wait for Font Awesome and BLKB Sans to be loaded.
-     */
-    waitForFaAndBbFonts(): Chainable<void>;
-
     /**
      * Capture a screenshot of the current page for visual regression testing.
      */
@@ -28,34 +11,6 @@ declare namespace Cypress {
     ): Chainable<void>;
   }
 }
-
-Cypress.Commands.add('waitForFonts', (...fonts: string[]) => {
-  return cy
-    .document()
-    .its('fonts.status', { timeout: 20000 })
-    .should('equal', 'loaded', { timeout: 20000 })
-    .end()
-    .document()
-    .then((doc) => {
-      cy.wrap(doc.fonts).should('not.be.undefined');
-      fonts.forEach((font) => {
-        cy.wrap(doc.fonts)
-          .invoke('check', `16px "${font}"`)
-          .should('be.true', undefined, { timeout: 20000 });
-      });
-    })
-    .end();
-});
-
-Cypress.Commands.add('waitForFontAwesome', () =>
-  cy.waitForFonts('FontAwesome')
-);
-Cypress.Commands.add('waitForBlackbaudSans', () =>
-  cy.waitForFonts('BLKB Sans')
-);
-Cypress.Commands.add('waitForFaAndBbFonts', () =>
-  cy.waitForFonts('BLKB Sans', 'FontAwesome')
-);
 
 /**
  * Capture a screenshot of the current page for visual regression testing. Include the URL in `blackout` because that is