+++ apps/e2e/ag-grid-storybook/src/app/ag-grid/ag-grid-stories.component.spec.ts
@@ -8,6 +8,8 @@ import {
 
 import { BehaviorSubject } from 'rxjs';
 
+import { FontLoadingTestingModule } from '../shared/font-loading/testing/font-loading-testing.module';
+
 import { AgGridStoriesComponent } from './ag-grid-stories.component';
 import { AgGridStoriesModule } from './ag-grid-stories.module';
 
@@ -29,7 +31,7 @@ describe('DataGridComponent', () => {
       }),
     };
     TestBed.configureTestingModule({
-      imports: [AgGridStoriesModule],
+      imports: [AgGridStoriesModule, FontLoadingTestingModule],
       providers: [
         {
           provide: SkyThemeService,