+++ apps/e2e/ag-grid-storybook/src/app/shared/font-loading/testing/font-loading-testing.module.ts
@@ -1,0 +1,15 @@
+import { NgModule } from '@angular/core';
+
+import { FontLoadingService } from '../font-loading.service';
+
+import { FontLoadingTestingService } from './font-loading-testing.service';
+
+@NgModule({
+  providers: [
+    {
+      provide: FontLoadingService,
+      useClass: FontLoadingTestingService,
+    },
+  ],
+})
+export class FontLoadingTestingModule {}