+++ apps/e2e/ag-grid-storybook/src/app/shared/font-loading/testing/font-loading-testing.service.ts
@@ -1,0 +1,12 @@
+import { Injectable } from '@angular/core';
+
+import { BehaviorSubject, Observable } from 'rxjs';
+
+@Injectable({
+  providedIn: 'root',
+})
+export class FontLoadingTestingService {
+  public ready(): Observable<boolean> {
+    return new BehaviorSubject(true);
+  }
+}