+++ apps/playground/src/app/components/popovers/dropdown/dropdown.component.ts
@@ -1,0 +1,20 @@
+import { Component } from '@angular/core';
+
+@Component({
+  selector: 'app-dropdown-demo',
+  templateUrl: './dropdown.component.html',
+})
+export class DropdownComponent {
+  // eslint-disable-next-line @typescript-eslint/no-explicit-any
+  public items: any[] = [
+    { name: 'Option 1', disabled: false },
+    { name: 'Disabled option', disabled: true },
+    { name: 'Option 3', disabled: false },
+    { name: 'Option 4', disabled: false },
+    { name: 'Option 5', disabled: false },
+  ];
+
+  public actionClicked(action: string): void {
+    alert(`You selected ${action}.`);
+  }
+}