+++ libs/components/ag-grid/src/lib/modules/ag-grid/cell-renderers/cell-renderer-lookup/cell-renderer-lookup.component.ts
@@ -20,8 +20,8 @@ import { SkyCellRendererLookupParams } from '../../types/cell-renderer-lookup-pa
 export class SkyAgGridCellRendererLookupComponent
   implements ICellRendererAngularComp
 {
-  public value = '';
-  public summaryCount = 0;
+  protected value = '';
+  protected summaryCount = 0;
 
   constructor(private changeDetector: ChangeDetectorRef) {}